% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr_nhanes.R
\name{check_formula_factors}
\alias{check_formula_factors}
\title{Check Formula Factors}
\usage{
check_formula_factors(formula, dataframe)
}
\arguments{
\item{formula}{A formula object.}

\item{dataframe}{A data frame containing the variables in the formula.}
}
\value{
A vector of factor variable names with fewer than two levels.
}
\description{
Identifies factor variables in a formula that have fewer than two levels.
}
\examples{
check_formula_factors(mpg ~ cyl + gear, mtcars)
}
