% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr_nhanes.R
\name{filter_na_columns}
\alias{filter_na_columns}
\title{Filter Columns with High NA Percentage}
\usage{
filter_na_columns(df, max_na_percent = 0.2)
}
\arguments{
\item{df}{A data frame.}

\item{max_na_percent}{Maximum allowed proportion of NA values per column. Default is 0.2.}
}
\value{
A data frame with filtered columns.
}
\description{
Filters out columns from a data frame that have a higher percentage of NA values than the specified threshold.
}
\examples{
filter_na_columns(data.frame(a = c(1, NA, 3), b = c(NA, NA, 3)), max_na_percent = 0.5)
}
