% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr_nhanes.R
\name{impute_and_report_na}
\alias{impute_and_report_na}
\title{Impute Missing Values and Report NA Counts}
\usage{
impute_and_report_na(mat)
}
\arguments{
\item{mat}{A numeric matrix.}
}
\value{
A matrix with imputed values.
}
\description{
This function imputes missing values in a matrix with column means and reports the count of missing values per column.
}
\examples{
mat <- matrix(c(1, NA, 3, 4, 5, NA), ncol = 2)
impute_and_report_na(mat)
}
