% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr_nhanes.R
\name{map_freq_to_numeric}
\alias{map_freq_to_numeric}
\title{Map Frequency Categories to Numeric Values}
\usage{
map_freq_to_numeric(df, colname)
}
\arguments{
\item{df}{A data frame containing the column to be mapped.}

\item{colname}{The name of the column to be mapped.}
}
\value{
A numeric vector with mapped values.
}
\description{
Maps specific frequency categories to numeric values, with error handling for invalid categories.
}
\examples{
map_freq_to_numeric(data.frame(freq = c("Not at all", "Several days")), "freq")
}
