% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr_nhanes.R
\name{matrix_from_latent}
\alias{matrix_from_latent}
\title{Generate a Matrix from a Latent Matrix with Added Noise}
\usage{
matrix_from_latent(latent_matrix, target_p, noise_sd = 0.3)
}
\arguments{
\item{latent_matrix}{A numeric matrix representing the latent structure (n x k).}

\item{target_p}{An integer specifying the number of columns in the output matrix.}

\item{noise_sd}{A numeric value indicating the standard deviation of the Gaussian noise to be added. Default is 0.3.}
}
\value{
A numeric matrix of dimensions (n x target_p), generated from the latent matrix
with a random linear transformation and added Gaussian noise.
}
\description{
This function generates a new matrix by applying a random linear transformation
to a given latent matrix and adding Gaussian noise.
}
\examples{
latent <- matrix(rnorm(20), nrow = 5, ncol = 4)
generated_matrix <- matrix_from_latent(latent, target_p = 3, noise_sd = 0.2)
print(generated_matrix)

}
