% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr_nhanes.R
\name{max_columns_list}
\alias{max_columns_list}
\title{Calculate the Maximum Number of Columns in a List of Matrices}
\usage{
max_columns_list(matrix_list)
}
\arguments{
\item{matrix_list}{A list containing matrices. Non-matrix elements will be ignored with a warning.}
}
\value{
An integer representing the maximum number of columns across all matrices in the list.
}
\description{
This function loops through a list of matrices and returns the maximum
number of columns found across all matrices in the list.
}
\examples{
mat1 <- matrix(1:6, nrow = 2)
mat2 <- matrix(1:12, nrow = 3)
mat3 <- matrix(1:4, nrow = 2)
max_columns_list(list(mat1, mat2, mat3)) # Returns 2

}
