% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr_nhanes.R
\docType{data}
\name{nhanes_dict}
\alias{nhanes_dict}
\title{NHANES data dictionary}
\format{
Curated nhanes cognition data dictionary from kaggle NGUYENVY
}
\usage{
nhanes_dict
}
\description{
NHANES data dictionary
}
\examples{
\dontrun{
 data("nhanes_dict")
}
}
\keyword{datasets}
