% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr_nhanes.R
\docType{data}
\name{nhanescog}
\alias{nhanescog}
\title{NHANES cognition data}
\format{
Curated nhanes cognition data from CERAD
}
\usage{
nhanescog
}
\description{
NHANES cognition data
}
\examples{
\dontrun{
 data("nhanescog_2011_2014")
}
}
\keyword{datasets}
