% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr_nhanes.R
\name{read_and_filter_xpt}
\alias{read_and_filter_xpt}
\title{Read and Filter XPT File}
\usage{
read_and_filter_xpt(file_path, threshold = 100)
}
\arguments{
\item{file_path}{Path to the XPT file.}

\item{threshold}{Numeric threshold to filter columns based on unique value counts. Default is 100.}
}
\value{
A data frame with filtered columns.
}
\description{
Reads an XPT file, converts it to a data frame, renames the first column to "ID", and filters based on unique value counts.
}
\examples{
read_and_filter_xpt("data.xpt", threshold = 50)
}
