% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr_nhanes.R
\name{remove_perfectly_correlated}
\alias{remove_perfectly_correlated}
\title{Remove Perfectly Correlated Columns from a Data Frame}
\usage{
remove_perfectly_correlated(df, tolerance = 1e-06)
}
\arguments{
\item{df}{A data frame containing numeric columns to be checked for perfect correlation.}

\item{tolerance}{float tolerance value}
}
\value{
A data frame with redundant perfectly correlated columns removed.
}
\description{
This function identifies columns in a data frame that are perfectly correlated
(correlation coefficient of 1 or -1) and removes the second instance from each
pair, ensuring that no two columns in the resulting data frame are perfectly correlated.
}
\examples{
df <- data.frame(a = 1:5, b = 1:5, c = 5:1+rnorm(5), d = c(2, 1, 6, 8, 10))
remove_perfectly_correlated(df)

}
